sign in and up done... can now be saved in the DB user can login with email or username

+++ source/src/main/java/artstore/entity/ArtPiece.java
@@ -15,8 +15,7 @@ public class ArtPiece {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long productId;
-
+    private int productId;
     private String title;
     private String description;
     private BigDecimal price;
@@ -27,11 +26,11 @@ public class ArtPiece {
     public ArtPiece() {
     }
 
-    public Long getProductId() {
+    public int getProductId() {
         return productId;
     }
 
-    public void setProductId(Long productId) {
+    public void setProductId(int productId) {
         this.productId = productId;
     }
